[CDEC-456] Reunite orphans from Pos.Recovery.Instance
