+++ infra/src/Pos/Infra/Recovery/Info.hs
@@ -6,6 +6,7 @@ module Pos.Infra.Recovery.Info
        , TipSlot (..)
        , SyncStatus (..)
        , MonadRecoveryInfo(..)
+       , MonadRecoveryInfoConstraints
        , recoveryInProgress
        , getSyncStatusK
        , recoveryCommGuard
@@ -14,12 +15,20 @@ module Pos.Infra.Recovery.Info
 
 import           Universum
 
+import qualified Control.Concurrent.STM as STM
+import           Control.Monad.Except (runExceptT, throwError)
 import           Formatting (bprint, build, sformat, stext, (%))
 import qualified Formatting.Buildable
 import           System.Wlog (WithLogger, logDebug)
 
-import           Pos.Core (HasProtocolConstants, SlotCount, SlotId, slotIdF,
+import           Pos.Core (HasProtocolConstants, SlotCount, SlotId,
+                     epochOrSlotG, epochOrSlotToSlot, flattenSlotId, slotIdF,
                      slotSecurityParam)
+import qualified Pos.DB.BlockIndex as DB
+import           Pos.DB.Class (MonadDBRead)
+import           Pos.Infra.Recovery.Types (RecoveryHeader, RecoveryHeaderTag)
+import           Pos.Infra.Slotting.Class (MonadSlots (getCurrentSlot))
+import           Pos.Util.Util (HasLens (..))
 
 newtype TipSlot = TipSlot SlotId
 
@@ -76,16 +85,61 @@ class Monad m => MonadRecoveryInfo m where
     -- are enumerated in 'SyncStatus'.
     getSyncStatus :: SlotCount -> m SyncStatus
 
+type MonadRecoveryInfoConstraints ctx m =
+    ( Monad m
+    , MonadIO m
+    , MonadDBRead m
+    , MonadSlots ctx m
+    , MonadReader ctx m
+    , HasLens RecoveryHeaderTag ctx RecoveryHeader
+    )
+
+instance ( Monad m
+         , MonadIO m
+         , MonadDBRead m
+         , MonadSlots ctx m
+         , MonadReader ctx m
+         , HasLens RecoveryHeaderTag ctx RecoveryHeader
+         , HasProtocolConstants
+         ) =>
+         MonadRecoveryInfo m where
+    getSyncStatus lagBehindParam =
+        fmap convertRes . runExceptT $ do
+            recoveryIsInProgress >>= \case
+                False -> pass
+                True -> throwError SSDoingRecovery
+            curSlotId <- note SSUnknownSlot =<< getCurrentSlot
+            tipHeader <- lift DB.getTipHeader
+            let curSlot = CurrentSlot curSlotId
+            let tipSlot@(TipSlot tipSlotId) = TipSlot $
+                    epochOrSlotToSlot (tipHeader ^. epochOrSlotG)
+            unless (tipSlotId <= curSlotId) $
+                throwError $
+                    SSInFuture tipSlot curSlot
+            let slotDiff = flattenSlotId curSlotId - flattenSlotId tipSlotId
+            unless (slotDiff < fromIntegral lagBehindParam) $
+                throwError $
+                    SSLagBehind tipSlot curSlot
+      where
+        recoveryIsInProgress = do
+            var <- view (lensOf @RecoveryHeaderTag)
+            isJust <$> atomically (STM.tryReadTMVar var)
+        convertRes :: Either SyncStatus () -> SyncStatus
+        convertRes (Left ss)  = ss
+        convertRes (Right ()) = SSKindaSynced
+
 -- | Returns if our 'SyncStatus' is 'SSDoingRecovery' (which is
 -- equivalent to “we're doing recovery”).
-recoveryInProgress :: MonadRecoveryInfo m => m Bool
+recoveryInProgress :: MonadRecoveryInfoConstraints ctx m => m Bool
 recoveryInProgress =
     getSyncStatus 0 {- 0 doesn't matter -} <&> \case
         SSDoingRecovery -> True
         _ -> False
 
 -- | Get sync status using K as lagBehind param.
-getSyncStatusK :: (MonadRecoveryInfo m, HasProtocolConstants) => m SyncStatus
+getSyncStatusK
+    :: MonadRecoveryInfoConstraints ctx m
+    => m SyncStatus
 getSyncStatusK = getSyncStatus lagBehindParam
   where
     -- It's actually questionable which value to use here. The less it
@@ -99,7 +153,7 @@ getSyncStatusK = getSyncStatus lagBehindParam
 -- kinda synchronized with the network.  It is useful for workers
 -- which shouldn't do anything while we are not synchronized.
 recoveryCommGuard
-    :: (MonadRecoveryInfo m, WithLogger m, HasProtocolConstants)
+    :: (MonadRecoveryInfoConstraints ctx m, WithLogger m)
     => Text -> m () -> m ()
 recoveryCommGuard actionName action =
     getSyncStatusK >>= \case
