+++ lib/cardano-sl.cabal
@@ -134,8 +134,6 @@ library
                         -- Network
                         Pos.Communication.Server
 
-                        Pos.Recovery.Instance
-
   build-depends:        base
                       , QuickCheck
                       , async
