+++ lib/src/Pos/Logic/Full.hs
@@ -36,6 +36,8 @@ import           Pos.DB.Ssc (sscIsDataUseful, sscProcessCertificate,
 import           Pos.DB.Txp.MemState (getMemPool, withTxpLocalData)
 import           Pos.DB.Update (getLocalProposalNVotes, getLocalVote,
                      isProposalNeeded, isVoteNeeded)
+import qualified Pos.GState.BlockExtra as DB (resolveForwardLink, streamBlocks)
+import           Pos.Infra.Recovery.Types (RecoveryHeader, RecoveryHeaderTag)
 import           Pos.Infra.Slotting (MonadSlots)
 import           Pos.Infra.Util.JsonLog.Events (JLEvent)
 import           Pos.Listener.Delegation (DlgListenerConstraint)
@@ -48,7 +50,6 @@ import qualified Pos.Network.Block.Logic as Block
 import           Pos.Network.Block.WorkMode (BlockWorkMode)
 import           Pos.Recovery (MonadRecoveryInfo)
 import qualified Pos.Recovery as Recovery
-import           Pos.Recovery.Types (RecoveryHeader, RecoveryHeaderTag)
 import           Pos.Security.Params (SecurityParams)
 import           Pos.Security.Util (shouldIgnorePkAddress)
 import           Pos.Ssc.Mem (sscRunLocalQuery)
