+++ lib/src/Pos/Recovery.hs
@@ -5,4 +5,3 @@ module Pos.Recovery
        ) where
 
 import           Pos.Infra.Recovery.Info
-import           Pos.Recovery.Instance ()
