+++ lib/src/Pos/Worker/Ssc.hs
@@ -47,7 +47,8 @@ import           Pos.DB.Ssc (getGlobalCerts, getStableCerts,
                      sscProcessOpening, sscProcessShares)
 import qualified Pos.DB.Ssc.SecretStorage as SS
 import           Pos.Infra.Diffusion.Types (Diffusion (..))
-import           Pos.Infra.Recovery.Info (MonadRecoveryInfo, recoveryCommGuard)
+import           Pos.Infra.Recovery.Info (MonadRecoveryInfoConstraints,
+                     recoveryCommGuard)
 import           Pos.Infra.Shutdown (HasShutdownContext)
 import           Pos.Infra.Slotting (MonadSlots, defaultOnNewSlotParams,
                      getCurrentSlot, getSlotStartEmpatically, onNewSlot)
@@ -83,7 +84,7 @@ type SscMode ctx m
       , MonadGState m
       , MonadDB m
       , MonadSscMem ctx m
-      , MonadRecoveryInfo m
+      , MonadRecoveryInfoConstraints ctx m
       , HasShutdownContext ctx
       , MonadReader ctx m
       , HasSscContext ctx
